backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
cycler==0.10.0
distlib==0.3.1
Django==3.1.6
djangorestframework==3.12.2
filelock==3.0.12
importlib-metadata==3.4.0
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
Pygments==2.7.4
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2021.1
six==1.13.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
virtualenv==20.4.2
zipp==3.4.0
